+++ packages/blocks-core/CHANGELOG.md
@@ -19,6 +19,7 @@
 - `Page List` block: `Current Page` inner block added.
 - `Gallery` block: improve support.
 - `Search` block: improve support.
+- `Post Navigation Link` block: improve support.
 
 
 ### Automated Tests